rename: BinningStrategies2 -> BinningStrategies
