+++ packages/@ourworldindata/grapher/src/color/BinningStrategies.ts
